+++ ujfe-signals/src/main/java/ujfe/signals/ComputedCycleException.java
@@ -1,0 +1,7 @@
+package ujfe.signals;
+
+public final class ComputedCycleException extends IllegalStateException {
+    public ComputedCycleException(String message) {
+        super(message);
+    }
+}