+++ ujfe-signals/src/main/java/ujfe/signals/SignalDependency.java
@@ -1,0 +1,5 @@
+package ujfe.signals;
+
+interface SignalDependency {
+    AutoCloseable subscribeInvalidation(Runnable listener);
+}